refactor: 응답 관련 파일명 response 로 변경
