+++ src/main/java/aurora/carevisionapiserver/domain/admin/exception/AdminException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.domain.admin.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class AdminException extends GeneralException {
