+++ src/main/java/aurora/carevisionapiserver/domain/bed/exception/BedException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.domain.bed.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class BedException extends GeneralException {
     public BedException(BaseErrorCode code) {
