+++ src/main/java/aurora/carevisionapiserver/domain/nurse/exception/NurseException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.domain.nurse.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class NurseException extends GeneralException {
     public NurseException(BaseErrorCode code) {
