+++ src/main/java/aurora/carevisionapiserver/domain/patient/exception/CameraException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.domain.patient.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class CameraException extends GeneralException {
     public CameraException(BaseErrorCode code) {
