+++ src/main/java/aurora/carevisionapiserver/global/auth/exception/AuthException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.global.auth.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class AuthException extends GeneralException {
     public AuthException(BaseErrorCode code) {
