+++ src/main/java/aurora/carevisionapiserver/global/fcm/exception/FcmException.java
@@ -1,7 +1,7 @@
 package aurora.carevisionapiserver.global.fcm.exception;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 
 public class FcmException extends GeneralException {
     public FcmException(BaseErrorCode code) {
