+++ src/main/java/aurora/carevisionapiserver/global/response/code/ReasonDTO.java
@@ -1,4 +1,4 @@
-package aurora.carevisionapiserver.global.error.code;
+package aurora.carevisionapiserver.global.response.code;
 
 import org.springframework.http.HttpStatus;
 
@@ -7,8 +7,9 @@ import lombok.Getter;
 
 @Getter
 @Builder
-public class ErrorReasonDTO {
-    private final HttpStatus httpStatus;
+public class ReasonDTO {
+
+    private HttpStatus httpStatus;
     private final boolean isSuccess;
     private final String code;
     private final String message;
