+++ src/main/java/aurora/carevisionapiserver/global/response/code/status/ErrorStatus.java
@@ -1,9 +1,9 @@
-package aurora.carevisionapiserver.global.error.code.status;
+package aurora.carevisionapiserver.global.response.code.status;
 
 import org.springframework.http.HttpStatus;
 
-import aurora.carevisionapiserver.global.error.code.BaseErrorCode;
-import aurora.carevisionapiserver.global.error.code.ErrorReasonDTO;
+import aurora.carevisionapiserver.global.response.code.BaseErrorCode;
+import aurora.carevisionapiserver.global.response.code.ErrorReasonDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -46,7 +46,10 @@ public enum ErrorStatus implements BaseErrorCode {
 
     // Camera
     CAMERA_NOT_FOUND(HttpStatus.NOT_FOUND, "CAMERA400", "카메라를 찾을 수 없습니다."),
-
+    EMPTY_FRAME(HttpStatus.BAD_REQUEST, "CAMERA401", "프레임이 비어있습니다."),
+    FRAME_PROCESSING_ERROR(HttpStatus.BAD_REQUEST, "CAMERA402", "프레임 처리 중 오류가 발생하였습니다."),
+    GRABBER_STOP_FAILED(HttpStatus.BAD_REQUEST, "CAMERA403", "스트리밍 중지에 실패하였습니다."),
+    EMPTY_S3_IMAGE(HttpStatus.NOT_FOUND, "CAMERA404", "S3에 이미지가 없습니다."),
     // Bed
     INVALID_BED_INFO(
             HttpStatus.BAD_REQUEST, "BED400", "잘못된 형식의 베드 정보입니다. '동 호 번' 또는 '호 번' 순서로 작성해 주세요."),
@@ -60,6 +63,8 @@ public enum ErrorStatus implements BaseErrorCode {
     AUTH_INVALID_TOKEN(
             HttpStatus.BAD_REQUEST, "FCM402", "FireStore에서 데이터를 불러오는 실행 도중 오류가 발생하였습니다."),
     UNSUPPORTED_TOKEN(HttpStatus.BAD_REQUEST, "FCM402", "FireStore에서 데이터를 불러오는 실행 도중 오류가 발생하였습니다."),
+    // FILE
+    FILE_CONVERT_FAIL(HttpStatus.BAD_REQUEST, "FILE400", "파일 변환에 실패하였습니다."),
     ;
     private final HttpStatus httpStatus;
     private final String code;
