+++ src/main/java/aurora/carevisionapiserver/global/security/handler/resolver/AuthUserArgumentResolver.java
@@ -14,8 +14,8 @@ import aurora.carevisionapiserver.domain.admin.domain.Admin;
 import aurora.carevisionapiserver.domain.admin.service.AdminService;
 import aurora.carevisionapiserver.domain.nurse.domain.Nurse;
 import aurora.carevisionapiserver.domain.nurse.service.NurseService;
-import aurora.carevisionapiserver.global.error.code.status.ErrorStatus;
-import aurora.carevisionapiserver.global.error.exception.GeneralException;
+import aurora.carevisionapiserver.global.response.code.status.ErrorStatus;
+import aurora.carevisionapiserver.global.response.exception.GeneralException;
 import aurora.carevisionapiserver.global.security.handler.annotation.AuthUser;
 import lombok.RequiredArgsConstructor;
