+++ src/main/java/aurora/carevisionapiserver/global/util/ApiExplorer.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import aurora.carevisionapiserver.domain.hospital.exception.HospitalException;
-import aurora.carevisionapiserver.global.error.code.status.ErrorStatus;
+import aurora.carevisionapiserver.global.response.code.status.ErrorStatus;
 
 @Component
 public class ApiExplorer {
