+++ src/main/java/aurora/carevisionapiserver/global/util/validation/validator/ExistNurseValidator.java
@@ -6,7 +6,7 @@ import jakarta.validation.ConstraintValidatorContext;
 import org.springframework.stereotype.Component;
 
 import aurora.carevisionapiserver.domain.nurse.service.NurseService;
-import aurora.carevisionapiserver.global.error.code.status.ErrorStatus;
+import aurora.carevisionapiserver.global.response.code.status.ErrorStatus;
 import aurora.carevisionapiserver.global.util.validation.annotation.ExistNurse;
 import lombok.RequiredArgsConstructor;
